+++ src/plugin/Components/CellCompiler.cs
@@ -61,9 +61,10 @@ namespace Sphere.Gdk.Components
 
         public async Task<bool> Package(IProject project, string fileName, IConsole con)
         {
-            string cellOptions = string.Format(@"--in-dir ""{0}"" --out-dir ""{0}/.staging"" --package ""{1}""",
+            var stagingPath = Path.Combine(project.RootPath, project.BuildPath);
+            string cellOptions = string.Format(@"--in-dir ""{0}"" --out-dir ""{2}"" --package ""{1}""",
                 project.RootPath.Replace(Path.DirectorySeparatorChar, '/'),
-                fileName.Replace(Path.DirectorySeparatorChar, '/'));
+                fileName.Replace(Path.DirectorySeparatorChar, '/'), stagingPath);
             if (m_main.Conf.MakeDebugPackages)
                 cellOptions += " --debug";
             return await RunCell(cellOptions, con);